Update macro-doc/1 and reinstate lib-docs
Fix #36.

Update macro-doc/1 to handle some not-so-edge-cases that came up
while running Lodox on lfex/clj. Mirror the form-doc/1 clauses for
symmetry and sanity.

N.B. defmacro clauses are wild beasts so be extra careful not to
consider a docstring a pattern/arglist. Also, macro clauses are not the
same as function clauses and can vary in arity, so handle them
(perhaps overly) explicitly.